Add license markers.
tomatosaver will be released under the ISC license; all contributions
and patches to tomatosaver must be released under the same license in
order to be considered for inclusion in the main repository.

The ISC license is functionally-equivalent to a 2-clause BSD license,
but it has the advantage of being written in plain, simple English.