Added countdown timer during breaks.
Added countdown timer during breaks to show exactly how much time
is left in the current break.  The animation is a little jerky and
the timer as implemented actually has an off-by-1 error in the way
it handles the last second because off the state change from the
break to waiting, but this is certainly good enough to get us
started. :)